+++ src/Payum/Paypal/ExpressCheckout/Nvp/Request/Api/DoVoid.php
@@ -1,0 +1,8 @@
+<?php
+namespace Payum\Paypal\ExpressCheckout\Nvp\Request\Api;
+
+use Payum\Core\Request\Generic;
+
+class DoVoid extends Generic
+{
+}