+++ src/Payum/Paypal/ExpressCheckout/Nvp/Tests/Request/Api/DoVoidTest.php
@@ -1,0 +1,18 @@
+<?php
+namespace Payum\Paypal\ExpressCheckout\Nvp\Tests\Request\Api;
+
+use Payum\Core\Request\Generic;
+use Payum\Paypal\ExpressCheckout\Nvp\Request\Api\DoVoid;
+
+class DoVoidTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * @test
+     */
+    public function shouldBeSubClassOfGeneric()
+    {
+        $rc = new \ReflectionClass(DoVoid::class);
+
+        $this->assertTrue($rc->isSubclassOf(Generic::class));
+    }
+}