+++ src/Payum/Skeleton/Action/CaptureAction.php
@@ -1,13 +1,16 @@
 <?php
 namespace Payum\Skeleton\Action;
 
-use Payum\Core\Action\GatewayAwareAction;
+use Payum\Core\Action\ActionInterface;
 use Payum\Core\Bridge\Spl\ArrayObject;
+use Payum\Core\GatewayAwareTrait;
 use Payum\Core\Request\Capture;
 use Payum\Core\Exception\RequestNotSupportedException;
 
-class CaptureAction extends GatewayAwareAction
+class CaptureAction implements ActionInterface
 {
+    use GatewayAwareTrait;
+
     /**
      * {@inheritDoc}
      *